requirements.txt
cnocr
clize
numpy
opencv-python==4.5.4.60
hanziconv
torchmetrics
torch
torchvision
transformers
jieba_fast
mxnet
flask
webvtt-py
Django
django-appconf
django-debug-toolbar
django-dotenv
gunicorn
jsonschema
zhon
wordfreq
sklearn
jieba_fast
jieba
pyperclip
sentence-transformers
iteration_utilities
word_forms
ckip_transformers
yt-dlp
